Refactor model structure: rename base model and update imports
