+++ pytorch_tabnet/tab_models/abstract_models/__init__.py
@@ -1,6 +1,6 @@
 """Abstract model definitions for TabNet."""
 
-from .base_model import TabModel
+from .abs_model import TabModel
 from .supervised_model import TabSupervisedModel
 
-__all__ = ["TabModel", "TabSupervisedModel"]
+__all__ = ["TabSupervisedModel", "TabModel"]
